Introduce Authenticatable
Use the Authenticatable union type to define
different methods of being authenticated.
Some types of Authenticatable are required
for some operations, for example deleting a
user always requires a password, so just a
token is not permitted.

Add a ToAuth trait to allow rocket input types
to be converted to Authentatable